refactor: reorganize image components and typography styles
- typography Extract font styles 新增 utils/fonts 檔案來封裝
- 修改 ImageWrapper 元件以符合新設計需求
- 新增具有響應式 CheckoutImage 元件
